+++ visualization/app/app.module.ts
@@ -1,38 +1,24 @@
-import "@angular/compiler" // todo this is needed for JIT compiler for ngColor within a downgraded component. Latest after full migration we can likely remove it again
-import "./app" // load AngularJS app first
+import "./app.scss"
+import "material-icons/iconfont/material-icons.css"
 import "zone.js" // needs to be loaded before "@angular/core"
-import { APP_INITIALIZER, Inject, NgModule } from "@angular/core"
+import { APP_INITIALIZER, NgModule, enableProdMode } from "@angular/core"
 import { BrowserModule } from "@angular/platform-browser"
 import { platformBrowserDynamic } from "@angular/platform-browser-dynamic"
-import { UpgradeModule } from "@angular/upgrade/static"
+import { HttpClientModule } from "@angular/common/http"
 import { FormsModule, ReactiveFormsModule } from "@angular/forms"
 import { MaterialModule } from "./material/material.module"
-import { AttributeSideBarModule } from "./codeCharta/ui/attributeSideBar/attributeSideBar.module"
-import { LegendPanelComponent } from "./codeCharta/ui/legendPanel/legendPanel.component"
-import { LegendPanelModule } from "./codeCharta/ui/legendPanel/legendPanel.module"
-import { ColorPickerForMapColorModule } from "./codeCharta/ui/colorPickerForMapColor/colorPickerForMapColor.module"
 import { EffectsModule } from "./codeCharta/state/angular-redux/effects/effects.module"
 import { UnfocusNodesEffect } from "./codeCharta/state/effects/unfocusNodes/unfocusNodes.effect"
 import { AddBlacklistItemsIfNotResultsInEmptyMapEffect } from "./codeCharta/state/effects/addBlacklistItemsIfNotResultsInEmptyMap/addBlacklistItemsIfNotResultsInEmptyMap.effect"
 import { dialogs } from "./codeCharta/ui/dialogs/dialogs"
-import { NodeContextMenuCardModule } from "./codeCharta/state/effects/nodeContextMenu/nodeContextMenuCard/nodeContextMenuCard.module"
 import { OpenNodeContextMenuEffect } from "./codeCharta/state/effects/nodeContextMenu/openNodeContextMenu.effect"
-import { FocusButtonsComponent } from "./codeCharta/state/effects/nodeContextMenu/focusButtons/focusButtons.component"
-import { LoadingFileProgressSpinnerModule } from "./codeCharta/ui/loadingFileProgressSpinner/loadingFileProgressSpinner.module"
-import { LoadingFileProgressSpinnerComponent } from "./codeCharta/ui/loadingFileProgressSpinner/loadingFileProgressSpinner.component"
 import { BlacklistSearchPatternEffect } from "./codeCharta/ui/searchPanel/searchBar/blacklistSearchPattern.effect"
-import { SliderModule } from "./codeCharta/ui/slider/slider.module"
 import { ResetColorRangeEffect } from "./codeCharta/state/store/dynamicSettings/colorRange/resetColorRange.effect"
-import { FileExtensionBarModule } from "./codeCharta/ui/fileExtensionBar/fileExtensionBar.module"
 import { ResetDynamicMarginEffect } from "./codeCharta/state/effects/resetDynamicMargin/resetDynamicMargin.effect"
-import { MetricChooserModule } from "./codeCharta/ui/metricChooser/metricChooser.module"
 import { ResetChosenMetricsEffect } from "./codeCharta/state/effects/resetChosenMetrics/resetChosenMetrics.effect"
-import { RibbonBarModule } from "./codeCharta/ui/ribbonBar/ribbonBar.module"
 import { UpdateEdgePreviewsEffect } from "./codeCharta/state/effects/updateEdgePreviews/updateEdgePreviews.effect"
 import { ChangelogDialogModule } from "./codeCharta/ui/dialogs/changelogDialog/changelogDialog.module"
 import { VersionService } from "./codeCharta/services/version/version.service"
-import { ActionIconModule } from "./codeCharta/ui/actionIcon/actionIcon.module"
-import { ToolBarModule } from "./codeCharta/ui/toolBar/toolBar.module"
 import { RenderCodeMapEffect } from "./codeCharta/state/effects/renderCodeMapEffect/renderCodeMap.effect"
 import { AutoFitCodeMapEffect } from "./codeCharta/state/effects/autoFitCodeMapChange/autoFitCodeMap.effect"
 import { CodeChartaModule } from "./codeCharta/codeCharta.module"
@@ -40,11 +26,13 @@ import { UpdateVisibleTopLabelsEffect } from "./codeCharta/state/effects/updateV
 import { ResetSelectedEdgeMetricWhenItDoesntExistAnymoreEffect } from "./codeCharta/state/effects/resetSelectedEdgeMetricWhenItDoesntExistAnymore/resetSelectedEdgeMetricWhenItDoesntExistAnymore.effect"
 import { LinkColorMetricToHeightMetricEffect } from "./codeCharta/state/effects/linkColorMetricToHeightMetric/linkColorMetricToHeightMetric.effect"
 import { UpdateFileSettingsEffect } from "./codeCharta/state/effects/updateFileSettings/updateFileSettings.effect"
+import { CodeChartaComponent } from "./codeCharta/codeCharta.component"
+import { NodeContextMenuCardModule } from "./codeCharta/state/effects/nodeContextMenu/nodeContextMenuCard/nodeContextMenuCard.module"
 
 @NgModule({
 	imports: [
 		BrowserModule,
-		UpgradeModule,
+		HttpClientModule,
 		EffectsModule.forRoot([
 			UnfocusNodesEffect,
 			AddBlacklistItemsIfNotResultsInEmptyMapEffect,
@@ -61,23 +49,12 @@ import { UpdateFileSettingsEffect } from "./codeCharta/state/effects/updateFileS
 			ResetSelectedEdgeMetricWhenItDoesntExistAnymoreEffect,
 			UpdateFileSettingsEffect
 		]),
-		SliderModule,
-		AttributeSideBarModule,
 		MaterialModule,
 		FormsModule,
-		LegendPanelModule,
-		ColorPickerForMapColorModule,
-		NodeContextMenuCardModule,
 		ReactiveFormsModule,
-		LoadingFileProgressSpinnerModule,
-		FileExtensionBarModule,
-		MetricChooserModule,
-		RibbonBarModule,
 		ChangelogDialogModule,
-		ActionIconModule,
-		ToolBarModule,
-		RibbonBarModule,
-		CodeChartaModule
+		CodeChartaModule,
+		NodeContextMenuCardModule
 	],
 	providers: [
 		VersionService,
@@ -89,14 +66,13 @@ import { UpdateFileSettingsEffect } from "./codeCharta/state/effects/updateFileS
 		}
 	],
 	declarations: [...dialogs],
-	entryComponents: [LegendPanelComponent, FocusButtonsComponent, LoadingFileProgressSpinnerComponent, ...dialogs]
+	entryComponents: [...dialogs],
+	bootstrap: [CodeChartaComponent]
 })
-export class AppModule {
-	constructor(@Inject(UpgradeModule) private upgrade: UpgradeModule) {}
+export class AppModule {}
 
-	ngDoBootstrap() {
-		this.upgrade.bootstrap(document.body, ["app"], { strictDi: true })
-	}
+if (process.env.DEV !== "true") {
+	enableProdMode()
 }
 
 platformBrowserDynamic().bootstrapModule(AppModule)