+++ visualization/app/codeCharta/ui/codeMap/rendering/geometryGenerator.spec.ts
@@ -1,25 +1,17 @@
 import { Node } from "../../../codeCharta.model"
-import { StoreService } from "../../../state/store.service"
+import { clone } from "../../../util/clone"
 import { STATE, TEST_NODE_ROOT } from "../../../util/dataMocks"
 import { BuildResult, GeometryGenerator } from "./geometryGenerator"
 
 describe("geometryGenerator", () => {
 	let geomGen: GeometryGenerator
-	let storeService: StoreService
 	let testNodes: Node[]
+	const state = clone(STATE)
+	state.dynamicSettings.heightMetric = "a" // set to a, since it is the delta defined in TEST_NODE_ROOT
 
 	beforeEach(() => {
-		mockStoreService()
 		initData()
 	})
-
-	const mockStoreService = () => {
-		STATE.dynamicSettings.heightMetric = "a" // set to a, since it is the delta defined in TEST_NODE_ROOT
-
-		storeService = jest.fn().mockReturnValue({
-			getState: jest.fn().mockReturnValue(STATE)
-		})()
-	}
 
 	const setTestNodes = () => {
 		const updatedNode = { heightDelta: -50 } // delta has to be negative why is that ?
@@ -43,7 +35,7 @@ describe("geometryGenerator", () => {
 
 		it("should add delta to height when not flattened", () => {
 			setFlattened(false)
-			buildResult = geomGen.build(testNodes, null, storeService.getState(), true)
+			buildResult = geomGen.build(testNodes, null, state, true)
 
 			expect(testNodes[0].flat).toBeFalsy()
 			expect(buildResult.desc.buildings[0].boundingBox.max.y).toBe(58)
@@ -51,7 +43,7 @@ describe("geometryGenerator", () => {
 
 		it("should not add delta to height when flattened", () => {
 			setFlattened(true)
-			buildResult = geomGen.build(testNodes, null, storeService.getState(), true)
+			buildResult = geomGen.build(testNodes, null, state, true)
 
 			expect(testNodes[0].flat).toBeTruthy()
 			expect(buildResult.desc.buildings[0].boundingBox.max.y).toBe(8)