+++ visualization/app/codeCharta/ui/customConfigs/customConfigHelper.service.ts
@@ -1,23 +1,22 @@
 import { combineLatest, map } from "rxjs"
-import { visibleFileStatesSelector } from "../../state/selectors/visibleFileStates.selector"
 import { CustomConfigHelper } from "../../util/customConfigHelper"
 import { getDownloadableCustomConfigs } from "./downloadCustomConfigsButton/getDownloadableCustomConfigs"
 import { Inject, Injectable } from "@angular/core"
 import { Store } from "../../state/angular-redux/store"
 import { getCustomConfigItemGroups } from "./customConfigList/getCustomConfigItemGroups"
-import { fileMapCheckSumsSelector } from "./customConfigList/fileMapCheckSums.selector"
+import { visibleFilesBySelectionModeSelector } from "./visibleFilesBySelectionMode.selector"
 
 @Injectable()
 export class CustomConfigHelperService {
 	readonly downloadableCustomConfigs$ = combineLatest([
-		this.store.select(visibleFileStatesSelector),
+		this.store.select(visibleFilesBySelectionModeSelector),
 		CustomConfigHelper.customConfigChange$
-	]).pipe(map(([visibleFileStates]) => getDownloadableCustomConfigs(visibleFileStates)))
+	]).pipe(map(([visibleFilesBySelectionMode]) => getDownloadableCustomConfigs(visibleFilesBySelectionMode)))
 
 	readonly customConfigItemGroups$ = combineLatest([
-		this.store.select(fileMapCheckSumsSelector),
+		this.store.select(visibleFilesBySelectionModeSelector),
 		CustomConfigHelper.customConfigChange$
-	]).pipe(map(([fileMapCheckSumsByMapSelectionMode]) => getCustomConfigItemGroups(fileMapCheckSumsByMapSelectionMode)))
+	]).pipe(map(([visibleFilesBySelectionMode]) => getCustomConfigItemGroups(visibleFilesBySelectionMode)))
 
 	constructor(@Inject(Store) private store: Store) {}
 }